TracBurndown 2.0.0dev: Fix incorrect types for milestone timestamp columns
Fixes #13257.


git-svn-id: https://trac-hacks.org/svn@16778 7322e99d-02ea-0310-aa39-e9a107903beb

+++ scrumburndownplugin/trunk/burndown/dbhelper.py
@@ -21,12 +21,13 @@ burndown_table = Table('burndown', key='id')[
     Column('hours_remaining', type='int'),
     Index(['id'])]
 
-milestone_table = Table('milestone', key='name')[Column('name'),
-                                                 Column('due', type='int'),
-                                                 Column('completed',
-                                                        type='int'),
-                                                 Column('description'),
-                                                 Column('started', type='int')]
+milestone_table = Table('milestone', key='name')[
+    Column('name'),
+    Column('due', type='int64'),
+    Column('completed', type='int64'),
+    Column('description'),
+    Column('started', type='int64')
+]
 
 
 def get_current_milestone(db, milestone_name):
